Reference Counting Garbage Collection
Introduce a new phase in the compiler that automatically inject proper reference counting method calls to manage the lifetime of objects. This change includes the naive implementation with some simple optimizations.

The reference counting garbage collection is currently turned off by default. To turn it on, enable the InitializeReferenceCountingGarbageCollection phase in the frontend config.

To opt out a specific type from being reference counted, use the DisableReferenceCounting attribute on the class.
To opt out a method from the automatic reference counting code injection, use the DisableAutomaticReferenceCounting attribute on the method.