[RefCountGC] Clean up the dummy bootstrap thread
During the initial stage of the bootstrap, we created a dummy thread so we can service ReleaseReference (which can call into ThreadImpl.CurrentThread.ReleaseReference) before a real thread object is there. This object, however, is previously being leaked. This change inject a snippet of code to clean it up by calling the memory manager directly just before the real thread object is assigned to ThreadImpl.CurrentThread.

Also refactor the one-time code injection pattern we have in ReferenceCountingGarbageCollection phase to avoid excessive duplicated code.