[RefCountGC] Make AddReference, ReleaseReference, LoadAndAddReference, Swap Native
Since the performance of AddReference, ReleaseReference, and other reference counting
helpers are critical in the performance of reference counting GC, move the
implementation to native to ensure that we only use the minimum number of instructions
needed to complete the operation.

The native code uses __set_PRIMASK(1) / __set_PRIMASK(0) to ensure the operation is
done atomically.

This also fixes an issue in previous LoadAndAddReference by making its operation atomic.
Without it, we can encounter memory corruption in the cases where multiple threads are reading
and writing to the same variable at the same time.

This improves the runtime performance by about 10%.