[RefCountGC] Skip reference counting on temporary used only to store the return value.
A typical pattern in zelig IR for returning an object type is:
   tmp = local;
   ....
   return tmp;

Without any optimization, we would inject AddRef / Release around the assignment and release the local at the end of the method, when in fact we can skip them all as long as local is not being redefined between the two operators.

This change added code in MethodTransformation to look for the above pattern, which results in elimination of around 400 AddRef / Release calls, and about 1.5K in code size / 4% increase in performance.