[RefCountGC] Adding Strict Reference Counting Garbage Collection
Adding an additional toggle to the strict mode for reference counting garbage collection (which is currently off by default), so we can further optimize the normal / non-strict mode by skipping reference counting on many variables that stores a local copy of a field / element / or indirect object.

In the strict mode, we assume all objects can be access from multiple thread concurrently. In the normal mode, we assume that an object will not work concurrently by default -- to access an object concurrently from multiple threads, appropriate mechanism, such as lock or volatile keyword need to be applied.

This allow us to cut about 2/3 of the total injections (from ~3K down to ~1K). And the code size and runtime overhead is 19KB and 30%.