[RefCountGC] Need to consider all non-argument variables in ref counting code injection
Previously, InjectReferenceCountingSetupAndCleanup is only considering TemporaryVariableExpression and LocalVariableExpression. With the introduction of Exceptions, we need to take care of those as well. So the new logic will apply ref count setup / cleanup on every variable that are not an ArgumentVariableExpression (arguments are special, they have implicit ref count from the caller).