Integrating LLVM opt and LLC into the compiler itself
- Added suport for spawning external processes opt.exe, llvm-dis, llc.exe into the zelig compiler directly. There are a few new command line options for controlling this behavior. "-GenerateObj" is the "outer most" triggering option. If it isn't present the compiler behaves as it always has thus far. By setting that option the new behavior is enabled. If "-SkipLlvmOpt" is also present, then the LLVM optimization pass is skipped. This is useful for testing/reviewing the unoptimized form. (Unless this is epclicityl set the efault behavior is to run the optimization passes), To allow further configuration and refinement of options the default arguments for opt.exe and LLC can be set via the "-LlvmOptArgs" and "-LlvmLlcArgs" options respectively.

One other command line option "-LlvmBinPath" is available to specify the location of the LLVM binary tools. If this is not set the code tries hard to find the tools based on our current usages (env vars from setenv/build.bat and the SDK properties files, registry entries from the LLVM build script etc...) or just the current PATH environment variable. If in the end it can't find the tools anywhere and the new behavior is enabled via "-GenerateObj" then an error is triggered informing the user about the problem without going any further.

- updated build.bat to default to skipping the opt and LLC calls, since now the default behavior is for that to occur from the compiler itself. They can be re-enabled if desired so it's not like that was removed.

NOTE:
This commit does NOT enable this support from the SDK build. I'll tackle updating the project files for that when we do the update of the project files etc... for the latest CPS bits from DevDiv
shortly.

- updated for review feedback
- added examples of command line options for LLVM code generation
- Fixed a bug in the compiler's command line parsing where it coudn't handle quoted string arguments with more than a single space