Fix incorrect return for AllocateFromManagedHeap and bug in AddReference
AllocateFromManagedHeap needs to return the pointer to the actual payload, since
it's being used by the interop codes to allocate memory from the managed heap --
it has no concept of ObjectHeader. Update both AllocateFromManagedHeap and
FreeFromManagedHeap to reflect this.

The code to transform an Object to ObjectHeader in AddReference was incorrect.