Remove the extra ReleaseReference injected when returning an argument that was modified.
With the current reference counting optimization, we only add and release reference on arguments that are modified within the method. Separately, we skip releasing the reference on the object that the method is returning. However, when a ref-counting argument is also the returning variable, we mistakenly perform an extra release.

Fixes #158.