CompilationOption for GarbageCollectionManager
Added a quick way to select between multiple implementations of GC available in the system.

To use, the platform (currently only LPC1768) needs to have multiple concrete implementations of GarbageCollectionManager (see BoardConfigurations\LPC1768\Board\SystemServices\GarbageCollectionManager.cs as an example). Once done, in the .FrontEndConfig, you can specify the GC strategy by using:

-CompilationOption System.String GarbageCollectionManager ConservativeMarkAndSweepCollector

(Replace the last bit with the name of the concrete implementation as needed)